Rename split-stacks-map into stacks-map
