+++ lib/migrate-new-resources.js
@@ -2,7 +2,7 @@
 
 const _ = require('lodash');
 
-const migrations = require("../split-stacks-map");
+const migrations = require("../stacks-map");
 
 module.exports = function migrateResources() {
   _.each(this.resourcesById, (resource, logicalId) => {
